# internal/dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Corral Dashboard API",
    "version": "2.0.0",
    "description": "The host-side control plane for Corral: repos, projects, GitHub issues, and the automations engine (actions, hooks, flows, runs, prompts) plus the host-wide logs and trace tree. The dashboard UI is one client; the `corral api` CLI and the host Claude skill are others. Loopback-only and token-gated. Mutating calls (POST/PUT/DELETE) require the operator to enable API writes in global settings; reads are always allowed."
  },
  "servers": [
    {
      "url": "/",
      "description": "Dashboard root (loopback-only, token-gated)"
    }
  ],
  "paths": {
    "/api/actions": {
      "get": {
        "summary": "List actions",
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Include this repo's actions plus global ones; omit for global only."
          }
        ],
        "responses": {
          "200": {
            "description": "Actions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "actions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Action"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an action",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Action"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "summary": "Fetch an action",
        "responses": {
          "200": {
            "description": "Action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Action"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update an action (name + spec)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "spec": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "delete": {
        "summary": "Delete an action",
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
    "/api/hooks": {
      "get": {
        "summary": "List hooks bound to an event",
        "parameters": [
          {
            "name": "event",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Event"
            }
          },
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hooks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Hook"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Bind an event to an action or flow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HookInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/hooks/{id}": {
      "delete": {
        "summary": "Unbind a hook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/api/flows": {
      "get": {
        "summary": "List flows",
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flows": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Flow"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a flow (optionally with inline steps)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "summary": "Fetch a flow with its steps",
        "responses": {
          "200": {
            "description": "Flow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a flow",
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
    "/api/runs": {
      "get": {
        "summary": "Recent run history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Run"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "summary": "Fetch a run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs": {
      "get": {
        "summary": "Application log (host-wide), newest-first, keyset-paginated",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "keyset cursor: rows older than this id"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "free-text search over message + meta"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LogRecord"
                      }
                    },
                    "nextCursor": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/prompts/project-start": {
      "get": {
        "summary": "Resolve the effective project-start prompt for a repo",
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Effective template + source + presets"
          }
        }
      }
    },
...
    "/api/prompts": {
      "get": {
        "summary": "Editable prompt catalog (default/effective/source/slots per prompt)",
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prompts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PromptItem"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{key}": {
      "parameters": [
        {
          "name": "key",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "repo",
          "in": "query",
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "summary": "Save a prompt override (repo scope with ?repo=)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "template": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated prompt item"
          }
        }
      },
      "delete": {
        "summary": "Reset a prompt (remove the override at this scope)",
        "responses": {
          "200": {
            "description": "Updated prompt item"
          }
        }
      }
    },
...
    "/repos": {
      "get": {
        "summary": "List repos (cross-project)",
        "responses": {
          "200": {
            "description": "Repos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "repos": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Repo"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a repo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repo"
                }
              }
            }
          }
        }
      }
    },
...
    "/repos/{id}": {
      "delete": {
        "summary": "Remove a repo (and its cache)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed"
          }
        }
      }
    },
...
    "/api/mcp": {
      "get": {
        "summary": "List connected MCP servers with live status",
        "description": "Drives `claude mcp list` on the host. status is connected | needs_auth | pending | unknown. A read \u2014 always allowed.",
        "responses": {
          "200": {
            "description": "Servers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "servers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/McpServer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Connect a remote MCP server",
        "description": "Runs `claude mcp add --scope user`. A mutating call \u2014 when made by the CLI/host Claude it requires API writes to be enabled; the browser is never gated.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpAddInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connected"
          },
          "403": {
            "description": "API writes disabled (CLI/Claude only)"
          }
        }
      }
    },
    "/api/mcp/{name}": {
      "delete": {
        "summary": "Remove an MCP server",
        "description": "Runs `claude mcp remove <name>`. Mutating \u2014 gated for the CLI/Claude, never for the browser.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed"
          },
          "403": {
            "description": "API writes disabled (CLI/Claude only)"
          }
        }
      }
    },
...
    "/api/prompts/library": {
      "get": {
        "summary": "List saved (named) prompts",
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prompts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NamedPrompt"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save a named prompt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamedPromptInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamedPrompt"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dind/caches/{name}": {
      "delete": {
        "summary": "Delete a DinD data cache",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Action": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/Kind"
          },
          "spec": {
            "type": "string",
            "description": "Per-kind config as a JSON string."
          },
          "scope": {
            "type": "string",
            "enum": [
              "global",
              "repo"
            ]
          },
          "repoId": {
            "type": "string"
          }
        }
      },
      "ActionInput": {
        "type": "object",
        "required": [
          "name",
          "kind"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/Kind"
          },
          "spec": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "global",
              "repo"
            ]
          },
          "repoId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}